// src/libs/interface.ts
export interface HospitalItem {
  _id: string;
  name: string;
  address: string;
  district: string;
  province: string;
  postalcode: string;
  tel: string;
  picture: string;
  __v: number;
  id: string;
}

export interface HospitalJson {
  success: boolean;
  count: number;
  pagination: object;
  data: HospitalItem[];
}